fix(ci): repair inherited main catalog, sequences, and typing failures
Keep one canonical sequences.core public path, type FiniteSequence
schema hooks as classmethods, rank finite-field irreducibility to
factor_mod_p, and move catalog boots out of the math test lane.
