+++ src/jacobian/math/number_theory/sequences/__init__.py
@@ -1,19 +1,5 @@
-"""Finite exact sequences and recurrence solving."""
+"""Finite exact sequences and recurrence solving.
 
-from jacobian.math.number_theory.sequences.core import (
-    AutocorrelationCell,
-    AutocorrelationResult,
-    FiniteIntegerSequence,
-    FiniteRationalSequence,
-    aperiodic_autocorrelation,
-    cyclic_autocorrelation,
-)
-
-__all__ = [
-    "AutocorrelationCell",
-    "AutocorrelationResult",
-    "FiniteIntegerSequence",
-    "FiniteRationalSequence",
-    "aperiodic_autocorrelation",
-    "cyclic_autocorrelation",
-]
+Canonical public sequence values and operations live in
+``jacobian.math.number_theory.sequences.core``.
+"""
